Add continue-on-error to create release step
Now that there are two of these running, one is going to win, and the
other's going to lose. We don't want the losing job to fall over just
for that reason.

However, I'm not sure if its output `upload_url` will be set, and we
need that later on, so we'll see what happens...